requirements.txt
asgiref==3.10.0
boto3==1.40.61
botocore==1.40.61
charset-normalizer==3.4.4
Django==4.2
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
mysqlclient==2.2.7
pillow==12.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reportlab==4.4.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
